[Driver: oo21696]
[Navigator av21927]

Added the timeToTakeMedicine() method. The times will be put in a csv file for each user. This will be found in the Medication App/src/resources/time. Also I now think "Time to Take (HH:mm):" prompt is not needed when a user wants to Add medicine. We should change this