[Driver:xl24773] [Navigator:mh22512]
1. Continue to adjust the structure, remove redundant files, specify the path of font files and icon paths
2. Preliminary repair of the permissions system, waiting for the completion of the relevant pages for further testing.
3. Partially reconstruct the existing gui interface.
4. Prepare to readjust the README document.

Warning: It is known that this modification will lead to the original About interface formatting errors, and will lead to the collapse of the drug management interface format, need to be partially repaired.

Note: This submission does not modify the command line version of the system, so it does not merge all the changes made to the email and medication reminders after 2024/11/25.