[Driver:xl24773]
1. Continue to adjust the structure, remove redundant files, adjust the file generation directory, using relative paths.
2. Serious error in fixing the permissions system.
3. At present, when you log in, it will automatically detect whether you have an account or not, and if you don't have an account, it will be created automatically by using the user name and password you inputted, so you don't need to jump to the second time.

Warning: It is known that this modification will lead to the original About interface formatting errors, and will lead to the collapse of the drug management interface format, need to be partially repaired.

Critical Warning: It is known that this change will cause the system to be unable to properly create accounts with permissions, users will not be assigned permissions, and the permissions tree system will crash completely, and attempts will be made to refactor some of the permissions tree code.

Note: This submission does not modify the command line version of the system, so it does not merge all the changes made to the email and medication reminders after 2024/11/25.